refactor sequence ids generator -> moved generator into sequence manager
